+++ src/main/java/carpet_client/CarpetClient.java
@@ -4,6 +4,7 @@ import carpet.CarpetExtension;
 import carpet.CarpetServer;
 import carpet_client.network.ServerMessageHandler;
 import carpet_client.utils.CarpetSettingsClientNetworkHandler;
+import carpet_client.utils.CarpetSettingsServerNetworkHandler;
 import carpet_client.utils.Reference;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
@@ -33,7 +34,7 @@ public class CarpetClient implements CarpetExtension
         CarpetServer.settingsManager.addRuleObserver((source, parsedRule, s) -> {
             try
             {
-                ServerMessageHandler.updateCarpetClientRules(parsedRule.name, parsedRule.getAsString(), source.getPlayer());
+                CarpetSettingsServerNetworkHandler.updateCarpetClientRules(parsedRule.name, parsedRule.getAsString(), source.getPlayer());
             }
             catch (CommandSyntaxException e)
             {
@@ -66,7 +67,7 @@ public class CarpetClient implements CarpetExtension
     @Override
     public void onPlayerLoggedIn(ServerPlayerEntity player)
     {
-        ServerMessageHandler.sendGUIInfo(player);
+        CarpetSettingsServerNetworkHandler.sendGUIInfo(player);
     }
 
     @Override
